server/requirements.txt
django
djangorestframework
rav
psycopg
drf_yasg
django-cors-headers
PyJWT
djangorestframework-simplejwt
cryptography